Move nginx-ingress addon logging configuration into nginx-ingress common folder
